--- practice/dual-contouring-terrain-main/.metaversefile
@@ -1,5 +1,0 @@
-{
-  "name": "terrain",
-  "description": "terrain",
-  "start_url": "index.js"
-}
