--- practice/dual-contouring-terrain-main/README.md
@@ -1,1 +1,0 @@
-# Generating Terrain
