--- practice/dual-contouring-terrain-main/biomes.json
@@ -1,342 +1,0 @@
-[
-  [
-    "biOcean",
-    6316128,
-    "Vol_26_5_Stone_"
-  ],
-  [
-    "biPlains",
-    9286496,
-    "Vol_11_4_Vegetation_"
-  ],
-  [
-    "biDesert",
-    16421912,
-    "Vol_16_4_Sand_"
-  ],
-  [
-    "biExtremeHills",
-    6316128,
-    "Vol_15_4_Rocks_"
-  ],
-  [
-    "biForest",
-    353825,
-    "Vol_17_1_Vegetation_"
-  ],
-  [
-    "biTaiga",
-    747097,
-    "Vol_21_1_Vegetation_"
-  ],
-  [
-    "biSwampland",
-    3145690,
-    "Vol_19_3_Vegetation_"
-  ],
-  [
-    "biRiver",
-    16440917,
-    "Vol_14_1_Dirt_"
-  ],
-  [
-    "biNether",
-    8323072,
-    "Vol_43_7_Lava_"
-  ],
-  [
-    "biEnd",
-    32767,
-    "Vol_42_4_Vegetation_"
-  ],
-  [
-    "biFrozenOcean",
-    9474192,
-    "Vol_22_1_Snow_Ice_"
-  ],
-  [
-    "biFrozenRiver",
-    16445632,
-    "Vol_22_4_Snow_Ice_"
-  ],
-  [
-    "biTundra",
-    16777215,
-    "Vol_20_4_Snow_Ice_"
-  ],
-  [
-    "biIceMountains",
-    10526880,
-    "Vol_20_3_Snow_Ice_"
-  ],
-  [
-    "biMushroomIsland",
-    16711935,
-    "Vol_43_1_Rocks_"
-  ],
-  [
-    "biMushroomShore",
-    10486015,
-    "Vol_43_5_Stone_"
-  ],
-  [
-    "biBeach",
-    16440917,
-    "Vol_16_4_Sand_"
-  ],
-  [
-    "biDesertHills",
-    13786898,
-    "Vol_16_6_Sand_"
-  ],
-  [
-    "biForestHills",
-    2250012,
-    "Vol_17_2_Vegetation_"
-  ],
-  [
-    "biTaigaHills",
-    1456435,
-    "Vol_21_4_Vegetation_"
-  ],
-  [
-    "biExtremeHillsEdge",
-    8359807,
-    "Vol_15_4_Rocks_"
-  ],
-  [
-    "biJungle",
-    5470985,
-    "Vol_75_4_Grass_"
-  ],
-  [
-    "biJungleHills",
-    2900485,
-    "Vol_75_1_Stone_"
-  ],
-  [
-    "biJungleEdge",
-    6458135,
-    "Vol_44_2_Stone_"
-  ],
-  [
-    "biDeepOcean",
-    3158064,
-    "Vol_26_5_Stone_"
-  ],
-  [
-    "biStoneBeach",
-    10658436,
-    "Vol_23_2_Stone_"
-  ],
-  [
-    "biColdBeach",
-    16445632,
-    "Vol_71_1_Stone_"
-  ],
-  [
-    "biBirchForest",
-    3175492,
-    "Vol_21_4_Vegetation_"
-  ],
-  [
-    "biBirchForestHills",
-    2055986,
-    "Vol_21_5_Vegetation_"
-  ],
-  [
-    "biRoofedForest",
-    4215066,
-    "Vol_21_1_Vegetation_"
-  ],
-  [
-    "biColdTaiga",
-    3233098,
-    "Vol_20_2_Snow_Ice_"
-  ],
-  [
-    "biColdTaigaHills",
-    5864818,
-    "Vol_21_1_Vegetation_"
-  ],
-  [
-    "biMegaTaiga",
-    5858897,
-    "Vol_19_2_Dirt_"
-  ],
-  [
-    "biMegaTaigaHills",
-    5858905,
-    "Vol_19_4_Dirt_"
-  ],
-  [
-    "biExtremeHillsPlus",
-    5271632,
-    "Vol_15_4_Rocks_"
-  ],
-  [
-    "biSavanna",
-    12431967,
-    "Vol_21_5_Vegetation_"
-  ],
-  [
-    "biSavannaPlateau",
-    10984804,
-    "Vol_19_5_Dirt_"
-  ],
-  [
-    "biMesa",
-    14238997,
-    "Vol_16_6_Sand_"
-  ],
-  [
-    "biMesaPlateauF",
-    11573093,
-    "Vol_23_6_Dirt_"
-  ],
-  [
-    "biMesaPlateau",
-    13274213,
-    "Vol_23_6_Dirt_"
-  ],
-  [
-    "biSunflowerPlains",
-    11918216,
-    "Vol_113_5_Flower_"
-  ],
-  [
-    "biDesertM",
-    16759872,
-    "Vol_40_4_Sand_"
-  ],
-  [
-    "biExtremeHillsM",
-    8947848,
-    "Vol_15_4_Rocks_"
-  ],
-  [
-    "biFlowerForest",
-    2985545,
-    "Vol_11_4_Vegetation_"
-  ],
-  [
-    "biTaigaM",
-    3378817,
-    "Vol_21_1_Vegetation_"
-  ],
-  [
-    "biSwamplandM",
-    522674,
-    "Vol_19_3_Vegetation_"
-  ],
-  [
-    "biIcePlainsSpikes",
-    11853020,
-    "Vol_36_1_Snow_Ice_"
-  ],
-  [
-    "biJungleM",
-    8102705,
-    "Vol_75_4_Grass_"
-  ],
-  [
-    "biJungleEdgeM",
-    6458135,
-    "Vol_44_2_Stone_"
-  ],
-  [
-    "biBirchForestM",
-    5807212,
-    "Vol_21_4_Vegetation_"
-  ],
-  [
-    "biBirchForestHillsM",
-    4687706,
-    "Vol_21_5_Vegetation_"
-  ],
-  [
-    "biRoofedForestM",
-    6846786,
-    "Vol_21_1_Vegetation_"
-  ],
-  [
-    "biColdTaigaM",
-    2375478,
-    "Vol_20_2_Snow_Ice_"
-  ],
-  [
-    "biMegaSpruceTaiga",
-    4542270,
-    "Vol_19_2_Dirt_"
-  ],
-  [
-    "biMegaSpruceTaigaHills",
-    4542286,
-    "Vol_19_4_Dirt_"
-  ],
-  [
-    "biExtremeHillsPlusM",
-    7903352,
-    "Vol_15_4_Rocks_"
-  ],
-  [
-    "biSavannaM",
-    15063687,
-    "Vol_21_5_Vegetation_"
-  ],
-  [
-    "biSavannaPlateauM",
-    10984820,
-    "Vol_19_5_Dirt_"
-  ],
-  [
-    "biMesaBryce",
-    16739645,
-    "Vol_27_3_Cliffs_"
-  ],
-  [
-    "biMesaPlateauFM",
-    14204813,
-    "Vol_23_6_Dirt_"
-  ],
-  [
-    "biMesaPlateauM",
-    15905933,
-    "Vol_23_6_Dirt_"
-  ],
-  [
-    "teDirt",
-    9268835,
-    "Vol_26_7_Rocks_"
-  ],
-  [
-    "teStone",
-    7901340,
-    "Vol_26_5_Stone_"
-  ],
-  [
-    "liWater",
-    4367861,
-    "Vol_36_3_Water_"
-  ],
-  [
-    "liWaterOcean",
-    16777215,
-    "Vol_36_2_Snow_Ice_"
-  ],
-  [
-    "liWaterRiverFrozen",
-    16777215,
-    "Vol_24_1_Snow_Ice_"
-  ],
-  [
-    "liWaterOceanFrozen",
-    16777215,
-    "Vol_24_2_Snow_Ice_"
-  ],
-  [
-    "liLava",
-    15684432,
-    "Vol_28_2_Lava_"
-  ]
-]
